Fixed bug when no annotations in database
Display of overlapping annotations was failing when the first annotation ended within a word shared by both it and the second (in display order) annotation.  The problem was a space inserted in the text retrieved for comparison between the recorded quote and the actual quote;  the space was incorrectly calculated for the EM element which, being inline, should not be considered to break words.  I don't know why I had commented out the code to check for an inline element, but I don't see it being called anywhere else so everything should be fine (I hope).


git-svn-id: http://marginalia.googlecode.com/svn/trunk/marginalia-lib@7 18b106f2-9c29-0410-bf68-a56e300ca647