Fixed some relatively rare errors in range calculations (which would have produced broken annotation locations in the database) at the cost of a probably significant performance hit.
There was a problem with how elements were counted for element paths.  The start
and end points for ranges are located relative to the nearest preceding tag for
a breaking element (where a breaking element is defined as one which would
correspond to whitespace or a break in a word - block, table cell, etc. elements
are breaking, but inline elements are not).  The problem is that for the rest
of the path ("parent" elements) were not calculated the same way.  In rare cases
this will be wrong (though I have never in fact encountered a problem);  in
particular, it could result in the existence of highlight ranges affecting
paths (so that when an annotation was added or deleted, subsequent annotations
might no longer be located correctly).

Take the following example document structure:
<ul>
<li>One <em>two</em> three
  <ul>
    <li>four</li>
  </ul>
</li>
</ul>

If "three" were highlighted, the path would be /1/1/3.0 (the zeroth character
in the third word following the first li tag) - not /1/1/1/2.0, because the
em tags are non-breaking so not counted (if they were p tags instead, the path
would indeed be /1/1/1/2.0).

Similarly, the path to "four" should be /1/1/1/1/1.0.  But under the old code,
the em tag *was* being counted since it was not the lowest element in the
path.  So the path would have been (incorrectly) /1/1/2/1/1.0.  The em
could of course be an annotation highlight, so the problem is immediately
apparent.  In test documents, however, (and probably on the web at large), it
is fairly rare for an an inline element to be a sibling of a block-level
element in content text (i.e. where annotations would be applied).

Why did I do something so stupid?  Well, it's because I wanted to use xpath
to locate elements.  The alternative is to do an O(n) count to locate the
nth element at each level - very slow.  Unfortunately, this is what I must
do, because I can't find a practical xpath expression for what I need,
because a) xpath doesn't have the concept of a breaking element, and b)
xpath makes it very difficult to match on multiple element names
(/(p|li|div|...)[1] doesn't work) and (so far as I can tell) impossible to
count only certain elements when indexing.

The result:  with no xpath, this new code will fix some edge cases at the
cost of performance.  It is increasingly imperative to find a faster way of
doing range calculations.


git-svn-id: http://marginalia.googlecode.com/svn/trunk/marginalia-lib@11 18b106f2-9c29-0410-bf68-a56e300ca647